ottimizzato accesso a database; sistemati controlli admin mancanti

+++ public/delete_prenotazione_user.php
@@ -3,14 +3,20 @@
 require_once "phplibs/databaseService.php";
 require_once "phplibs/templatingService.php";
 
-$database = new DatabaseService();
-//user param sarà da rimpiazzare con session user_id implementate le funzionalità di login
+if (!isset($_SESSION['user_id'])) {
+    header("Location: login.php");
+    exit;
+}
 
-$alterSuccess = $database->deletePrenotazioneUser($_SESSION['user_id'], $_POST['data_prenotazione']);
+try {
+    $database = new DatabaseService();
+    $alterSuccess = $database->deletePrenotazioneUser($_SESSION['user_id'], $_POST['data_prenotazione']);
+    unset($database);
+} catch (Exception $e) {
+    unset($database);
+    Templating::errCode(500);
+}
 
-if ($alterSuccess)
-    header('Location: profile.php');
-else
-    header('Location: index.php');
+header('Location: profile.php');
 exit;
 ?>
